First pass at openid support.
Works the similar to the others, but a endpoint url is required, in
this form:

/auth/openid/?openid_identifier=rmorison.pip.verisignlabs.com

where openid_identifier is the endpoint, with or without http(s),
suitable encoded of course.

There are open issues about user naming, expire dates, and a few other
things. Needs test against more servers, too.